ifupdown-migrate: operate on /e/n/i in-order
/etc/network/interfaces should be operated on in-order. Using a plain
dict as the data store for parsing it does not provide this guarantee.
This can lead to unexpected results when later parts of the file are
parsed before earlier parts.

Use OrderedDicts to ensure that interfaces and families are parsed
in the order they are provided.
